+++ src/nvim/mbyte.c
@@ -3956,8 +3956,8 @@ static int check_col(int col)
 {
   if (col < 0)
     return 0;
-  if (col >= (int)screen_Columns)
-    return (int)screen_Columns - 1;
+  if (col >= screen_Columns)
+    return screen_Columns - 1;
   return col;
 }
 
@@ -3966,7 +3966,7 @@ static int check_row(int row)
 {
   if (row < 0)
     return 0;
-  if (row >= (int)screen_Rows)
-    return (int)screen_Rows - 1;
+  if (row >= screen_Rows)
+    return screen_Rows - 1;
   return row;
 }
